switch dbc records to NamedStruct
format was passed to DbcFile constructor as a (<format string>, [<string fields>,]) pair. instead of a format string, the first element is a NamedStruct created with the names and field types from the format definition.

methods added to NamedStruct to make it behave exactly like its struct.Struct member.

in the DbcFile.load method, some fuckery was added in order to add strings to the loaded records, which was previously done by creating a list from the tuple given from struct.unpack, for mutability, and replacing the string offsets with the looked up string in place. now, DbcFile.load creates a dictionary of the string fields and merges it into a dictionary of the entire record, then reaches into the NamedStruct instance to create a new instance of its underlying namedtuple with the contents of the merged dictionary.